on the way to removing Joda Time and javax.measure

+++ com.github.rasifix.trainings.core/src/main/java/com/github/rasifix/trainings/model/Trackpoint.java
@@ -38,7 +38,7 @@ public class Trackpoint {
 	
 	private Measure<Duration> elapsedTime;
 
-	public Trackpoint(final double elapsedTime) {
+	public Trackpoint(final long elapsedTime) {
 		this(Measure.valueOf(elapsedTime, SI.MILLI(SI.SECOND)));
 	}
